cloud-run-ocr/requirements.txt
flask==3.0.0
google-cloud-documentai==2.24.0
gunicorn==21.2.0
